requirements.txt
numpy==1.24.2
python-dotenv==0.21.0
tensorflow==2.10
sumolib==1.16.0
eclipse-sumo==1.16.0
traci==1.16.0
pydot==1.4.2
matplotlib==3.7.1
